Rename DelegatioEdge to Delegation
