+++ packages/api/src/fns/delegations/createDelegations.spec.ts
@@ -1,9 +1,9 @@
 import { describe, test } from '@jest/globals'
 import { Address } from 'viem'
 
-import createEdges from './createEdges'
+import createDelegations from './createDelegations'
 
-describe('createEdges', () => {
+describe('createDelegations', () => {
   const A = 'A' as Address
   const B = 'B' as Address
   const C = 'C' as Address
@@ -16,7 +16,7 @@ describe('createEdges', () => {
         optOut: false,
       },
     }
-    const result = createEdges(registry, 0)
+    const result = createDelegations(registry, 0)
 
     expect(result).toEqual([
       { delegator: A, delegate: B, weight: 100, expiration: 0 },
@@ -34,7 +34,7 @@ describe('createEdges', () => {
         optOut: false,
       },
     }
-    const result = createEdges(registry, 0)
+    const result = createDelegations(registry, 0)
 
     expect(result).toEqual([
       { delegator: A, delegate: A, weight: 20, expiration: 0 },
@@ -58,7 +58,7 @@ describe('createEdges', () => {
         optOut: true,
       },
     }
-    const result = createEdges(registry, 0)
+    const result = createDelegations(registry, 0)
 
     expect(result).toEqual([
       { delegator: A, delegate: B, weight: 50, expiration: 0 },
@@ -81,7 +81,7 @@ describe('createEdges', () => {
         optOut: false,
       },
     }
-    const result = createEdges(registry, 2024)
+    const result = createDelegations(registry, 2024)
 
     expect(result).toEqual([
       { delegator: B, delegate: C, weight: 100, expiration: 2025 },
@@ -104,7 +104,7 @@ describe('createEdges', () => {
         optOut: true,
       },
     }
-    const result = createEdges(registry, 2024)
+    const result = createDelegations(registry, 2024)
 
     expect(result).toEqual([
       { delegator: A, delegate: C, weight: 50, expiration: 0 },
@@ -124,7 +124,7 @@ describe('createEdges', () => {
         optOut: true,
       },
     }
-    const result = createEdges(registry, 2024)
+    const result = createDelegations(registry, 2024)
 
     expect(result).toEqual([
       { delegator: B, delegate: C, weight: 100, expiration: 0 },
