+++ packages/api/src/fns/delegations/createGraph.ts
@@ -2,29 +2,29 @@ import assert from 'assert'
 
 import findCycle from '../graph/findCycle'
 
-import { DelegationEdge } from './types'
+import { Delegation } from './types'
 import { Graph } from '../../types'
 
 export default function createGraph(
-  edges: DelegationEdge[]
+  delegations: Delegation[]
 ): Graph<{ expiration: number; weight: number }> {
   const result: Graph<{ expiration: number; weight: number }> = {}
-  for (const { delegator, delegate, weight, expiration } of edges) {
+  for (const { delegator, delegate, weight, expiration } of delegations) {
     result[delegator] = result[delegator] || {}
     result[delegator][delegate] = { weight, expiration }
   }
 
   const cycle = findCycle(result)
-  return cycle ? createGraph(filterOldestCycleEdge(edges, cycle)) : result
+  return cycle ? createGraph(filterOldestCycleEdge(delegations, cycle)) : result
 }
 
 /*
  * Idea is that the edge that created the cycle is the one that gets busted
  */
 function filterOldestCycleEdge(
-  edges: DelegationEdge[],
+  edges: Delegation[],
   cycle: [string, string][]
-): DelegationEdge[] {
+): Delegation[] {
   const entries = cycle
     .map(([from, to]) =>
       edges.findIndex(
